add log_syscall usage to handle_syscall

+++ syscall_format.cpp
@@ -5,7 +5,7 @@
 #include <unordered_map>
 
 namespace {
-    const std::unordered_map<int, std::pair<std::string, std::string>> sys_map = {
+    const std::unordered_map<unsigned long long, std::pair<std::string, std::string>> sys_map = {
         {0, {"read", "data read"}},
         {1, {"write", "data output"}},
         {2, {"open", "file access"}},
@@ -20,7 +20,7 @@ namespace {
  * @param num the syscall number
  * @return string explaining which syscall was used
  */
-std::pair<std::string, std::string> get_syscall_name(const int num) {
+std::pair<std::string, std::string> get_syscall_name(const unsigned long long num) {
     auto it = sys_map.find(num);
     if (it != sys_map.end()) {  //syscall number found in sys_map
         return it->second;
